requirements.txt
bcrypt==4.0.1
click==8.1.3
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2==2.9.5
PyJWT==2.6.0
python-dotenv==0.21.1
SQLAlchemy==2.0.4
typing_extensions==4.5.0
Werkzeug==2.2.3
